+++ ReleaseNotes.md
@@ -1,3 +1,26 @@
+# 0.9.0 (17 April 2026)
+- [#19](https://github.com/StefH/JsonConverter/pull/19) - JsonConverterOptions:DateParseHandling [Feature] contributed by [StefH](https://github.com/StefH)
+
+# 0.8.0 (20 December 2025)
+- [#17](https://github.com/StefH/JsonConverter/pull/17) - Fix IgnoreNullValues for NewtonsoftJsonConverter [bug] contributed by [StefH](https://github.com/StefH)
+- [#18](https://github.com/StefH/JsonConverter/pull/18) - .NET 10 [Feature] contributed by [StefH](https://github.com/StefH)
+
+# 0.7.2 (13 June 2025)
+- [#16](https://github.com/StefH/JsonConverter/pull/16) - Newtonsoft.Json: support incompatible types in array (merge to common structure) [Feature] contributed by [StefH](https://github.com/StefH)
+
+# 0.7.1 (16 November 2024)
+- [#15](https://github.com/StefH/JsonConverter/pull/15) - Update JsonConverter.System.Text.Json to .NET 9 [Feature] contributed by [StefH](https://github.com/StefH)
+
+# 0.7.0 (10 October 2024)
+- [#13](https://github.com/StefH/JsonConverter/pull/13) - Bump System.Text.Json from 6.0.6 to 6.0.10 in /src/JsonConverter.System.Text.Json [dependencies] contributed by [dependabot[bot]](https://github.com/apps/dependabot)
+- [#14](https://github.com/StefH/JsonConverter/pull/14) - Bump System.Text.Json from 8.0.4 to 8.0.5 + upgrade some other packages [Feature] contributed by [StefH](https://github.com/StefH)
+
+# 0.6.0 (13 July 2024)
+- [#12](https://github.com/StefH/JsonConverter/pull/12) - Add .NET 8 support to JsonConverter.System.Text.Json [Feature] contributed by [StefH](https://github.com/StefH)
+
+# 0.5.0 (21 December 2023)
+- [#11](https://github.com/StefH/JsonConverter/pull/11) - Add &quot;object? Deserialize(string text, Type type, ...)&quot; method [Feature] contributed by [StefH](https://github.com/StefH)
+
 # 0.4.0 (02 March 2023)
 - [#10](https://github.com/StefH/JsonConverter/pull/10) - Add ConvertToDynamicJsonClass and DeserializeToDynamicJsonClass [Feature] contributed by [StefH](https://github.com/StefH)